requirements.txt
requests==2.18.4
python-telegram-bot==10.0.1